Move nativation to different component
