+++ WBooks/src/app/components/Navigation/styles.js
@@ -1,0 +1,10 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  tabBarItem: {
+    width: 25,
+    height: 25
+  }
+});
+
+export default styles;
